The auto generated code now works with the GUI runner. This means it should also work with the NUI as well.
I had to change a few things to get this to work. Now some of the
functions take inn the msp-object, but that works out fine as long as
it is remembered to add it to the  end of the list as the expand
function reverses the sequence of the function.