CountDownLatch added and working for NUIMsPacman
NUIMsPacman now keeps a CountDownLatch to ensure that the bot can
start. This provides a much quicker, and most importantly, SAFER start.

Can not do the same for GUIMsPacman because of the frame running in the
same thread as the the CountDownLatch awaits the countdown to 0.

Removed unused code from core.

Put population down to 50 because of a VERY long runtime with 70.

Removed CountDownLatch from GUIMsPacman because of the reasons given.

Made fitness-graphics run the same way (a loop) as fitness.