I think I found the bug in fitness
We need one more countdown latch then there are tries
    The reason for this is that stopMSP might run after msp has gone
through the next round.

When trying to finish the priority of the game thread is put to max.
   So that we are sure that we finish when we can.

Join no longer times out.
    We should not experience hangups here anymore, if we do then we
need to create an exception so it fails.

Removed the let around sc
   Really don't need this anymore

Changed name for variables,
   Changed to more understandable names