Moved to a new idea/refactored
The general way of running the game through the mspacman-classes has
been refactored.
Now there is a game class that runs the game and the different mspacman
classes are responsible are only responsible for GUI and input.

There is a new Experimental class.
The new exp. is there to test the new MAP that has been created. Now we
can tell the path between mspacman and a point on the map. We use A* to
do that.